functions/publishers/twitter-publisher/requirements.txt
google-cloud-secret-manager==1.0.*
python-twitter==3.5.*
